backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1
celery==5.4.0
redis==5.0.4
boto3==1.34.110
python-multipart==0.0.9
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
httpx==0.27.0
Pillow==10.3.0
numpy==1.26.4
opencv-python-headless==4.9.0.80
torch==2.3.0
torchvision==0.18.0
diffusers==0.27.2
transformers==4.41.0
accelerate==0.30.0
peft==0.11.0
safetensors==0.4.3
scipy==1.13.0
scikit-learn==1.4.2
